Fixed some bugs

+++ PaddleRS/login/objectdetection.py
@@ -120,7 +120,7 @@ def recvImg(request):
         threshold=0.2, save_dir=None
     )
     print(cname,score)
-    out = cv2.cvtColor(out,cv2.COLOR_BGR2RGB)
+    out = cv2.cvtColor(out.astype(np.uint8),cv2.COLOR_BGR2RGB)
     cv2.imwrite('od_out.png',out)
   retval,img_buffer = cv2.imencode('.jpg', out)
   return JsonResponse({'img':'data:false;base64,'+str(base64.b64encode(img_buffer))[2:-1],'list':{'id':1,'class':cname,'score':score}})
